Don't attempt to create the sound device
If /dev/dsp is missing, don't create it.  Otherwise you end up with a
file containing all the audio produced during execution.

Now it is possible that someone wants to capture the audio to a file,
but the code performs ioctl() on the file descriptor which makes no
sense for a regular file.  We won't cater to that use case.