Simplify Game Genie decoding
After some observations, most of the ugly bit manipulation is to rotate
the high bit from each code letter to the next code letter.  Do this in
one step, then rearrange the nibbles.

The high bit of the 2nd code letter is unused.  It would serve as
address bit 15, but we OR the address with 0x8000.