Switch scaler to use XRender extension; remove scanline customization (-L option)
Here we change the semantics of the framebuffer to be the raw image, with no
filter applied.  We use XRender to scale instead of line doubling as each frame
is rendered.  This may pave the way for more advanced filters later on.

But to do this we also need to remove the -L option / the "scanlines" variable,
so scanline customization is ripped out.