Isolate HQX build steps from the rest of tuxnes
Ideally we should be using the system HQX library, but it appears that
there are few distros that currently package it.  Prefer the system
version by default; build our own if missing.

When building our own, ideally we should use autoconf nested packages
(see AC_CONFIG_SUBDIRS) but that comes with its own problems, notably:
* Builds (and installs, during `make install`) the whole HQX project
  when we just want the library,
* Builds the library in both static and shared configurations as a
  libtool library, when we just want the static lib,
* Linking to the libtool library will prefer the shared lib which we
  don't want to install since it could conflict with a system lib,
* The static lib is hidden in submodules/hqx/.libs/libhqx.a which is
  probably a fragile path to depend upon, and there is no specific
  target to build only the static lib.

Since we have to mangle the user-specified CFLAGS to force -O3, I've
moved the HQX library to a different Makefile.am so that mangling won't
apply to the tuxnes sources.  It doesn't necessarily need to be in the
toplevel Makefile.am, it was just convenient to put it there.