Simplify handling of long options with no short equivalents
getopt_long can return values outside of the ASCII range for anything
without a short equivalent.  Avoids the need to do table index
comparisons, which was made more complicated here by doing an additional
string comparison.

Also we don't want getopt_long to return 0 for this case because that
value is reserved for when the 3rd table column has a non-NULL value.