Try fixing the basename parsing again
What we want to do is to strip everything before and including the last
'/', then everything after and including the last '.', but the last
three attempts all had problems computing the length.  If the length is
negative, then the malloc could fail.  Here's another attempt.

Until 2da25733c46d92a7d89b6ad74de65c9445e72187:  If the last '.'
precedes the last '/', then the length will be negative.

Until 53cfdd22dcf05ff346fa5e3f39f3bf399ca8dfcc:  The basename length is
reduced by the length of the extension (including the '.').  But if
there are multiple '.' characters, this process happens more than once,
which could make the length go negative.  I may have meant to break
after the first '.', but that behavior wouldn't be right either.

Until now:  The basename length is taken as the last extension length
(including the '.') if any extension is present.  It will never be
negative, but is unlikely to be correct.

Also none of these attempts did a good job of handling filenames that
begin with '.' and have no other extension; it would result in an empty
string.  Or if the filename begins with multiple '.' characters, it
could produce something undesirable like "." or ".." (likely to be
interpreted as a directory).  So to avoid these, skip over any '.'
characters at the start of a directory component.

Other thoughts:
* It might be cleaner to do the parsing with strchr and strrchr, but
  this loop does it in a single pass over the string.
* A filename of "foo.nes.gz" gets converted to "foo.nes".  Maybe it
  makes sense to strip everything after the first '.', or maybe we
  shouldn't try stripping extensions.  The basename is currently only
  used for screenshots.