Change how compdata is inserted into the executable
The goal is to put it into the .rodata section (to actually make it
read-only) and to 8-byte-align it (to match how comptbl.c inserts
instructions), but the existing $(LD) command puts it into .data which
is writable, and does not specify any alignment constraints.  It also
adds symbols _binary_compdata_(start|end|size) as globals that we don't
want or use, and it is not setting the symbol types or sizes correctly.

Most (but not all) of these issues could be fixed using `objcopy`, but
the commands might require a particular version of it or otherwise be
fragile.

The assembler .incbin directive seems to be the most portable way to
include binary data, and it allows fixing all other issues noted above.
Also since compdata is all x86-specific at the moment, it seems to
belong here.