Fix several bytes-per-line calculations
fb.c/pixels.h:  Remove nextline variable as an unnecessary conversion
from bytes to pixel_t for the pointer arithmetic.  The conversion is
wrong if bytes_per_line is not an integer multiple of sizeof (pixel_t).

x11.c:  Always create the XImage first, then calculate the required
buffer size as image->bytes_per_line * image->height.  The previous
calculation for shmget() used an uninitialized (0) value for
bytes_per_line, causing this call to fail.  The previous calculation for
the unscaled image buffer had an extra divide-by-8 causing it to be too
small.

When the last argument to XCreateImage (bytes_per_line) is 0,
image->bytes_per_line is automatically calculated as (width * bpp),
rounded up to the next bitmap_pad boundary, and converted from bits to
bytes.

bitmap_pad is the next-to-last argument and must be 8, 16, or 32 (in
bits).  This is meant for any additional padding at the end of a
scanline so that the next scanline starts on a particular bitmap_pad
boundary, say if pixels are packed 24bpp, scanlines must start on a
4-byte boundary, and width isn't a multiple of 4.  Set it to the
display's default bitmap_pad.  Using bpp for this argument fails unless
it happens to be one of these three acceptable values.  Our image
drawing code does not care about the bitmap_pad value since it always
advances by bytes_per_line.

HQX needs bpp == 32, not depth == 32.  It also makes assumptions about
the pixel format that are not always valid (particularly for palettized
displays).