Reconcile --geometry with minimum window size
--geometry does not apply when --in-root is specified.  Ignore it unless
we are creating a window, in which case the user-specified x/y/w/h needs
to be used for both the XCreateWindow() call and the size hints for it
to appear in the correct location and at the correct size.

XWMGeometry (via XParseGeometry) properly handles a null geometry
string, and the returned w/h will be at least the minimum w/h (which we
were previously not enforcing).  The base w/h is set to 0 or else it
defaults to the minimum w/h which causes the window to end up larger
than the user specified.