Fix shared memory leaks
After a successful shmget(), we must clean up using

    shmctl(shminfo.shmid, IPC_RMID, NULL)

or else the shared memory segment sticks around forever, even after all
processes have exited.  It was fork()ing and expecting the parent to do
the cleanup, but that was wasteful and had several problems:

* shm_attached and shminfo.shmaddr were never valid in the parent,
  so the cleanup handler never performed those parts.  Additionally
  shmat() sets shminfo.shmaddr to -1, not 0, on error, so the shmdt()
  portion of the cleanup was not guarded appropriately.
* Signals can cause it to exit without running the cleanup handler.

The safe, but not necessarily portable way to ensure we don't leak the
segment is to do the IPC_RMID just after all processes have attached.
Most Unices support this, but some may delete the share immediately
instead of waiting until everyone detaches.  Linux allows attaching
after the IPC_RMID, but that's even less portable so I won't attempt it.

It is still vulnerable to leaks if the process dies between the shmget()
and IPC_RMID, but should be much better now.

Also restrict permissions on the shared memory segment from 0777 to
0600.  There is enough info in /proc/sysvipc/shm to allow anyone to
attach and read/write the framebuffer memory with the previous perms.