Fix issues with palette argument parsing
The -p and -P options both have a required argument, but the handling
for both seem to treat it as an optional argument.  '-P ""' was handled
as '-P loopy' (where loopy is the default palette), is now handled as an
error instead.

Isolate the table lookup from the palette assignment somewhat.  palfile
is cleared when NES_palette is set, and vice-versa, except for the -N
handling for some reason.