Update whitespace handling in iNES palette parser
Don't require whitespace after the very first pen, in case there is only
one pen.  Do require whitespace between pens, enforced by checking that
count == 1 inside the loop which indicates that we've done at least one
memmove and successful byte read.  Whitespace after the last pen is (and
has been) optional.

Avoid an infinite loop if the file ends with 8+ bytes of consecutive
whitespace.  The memmove never shifts any non-whitespace characters into
a position to terminate the loop, and a read of 0 once the end-of-file
is reached did not terminate the loop either.

Let isspace() check for '\r' or '\n'.  Beware, all of these ctype
functions are unnecessarily locale-dependent and should be replaced.

The bottom half of the loop now pre-loads buf with the next pen.