Speed up and simplify NTSC palette generation
Use a single loop instead of two nested loops.  Extract x and z from the
loop index.

Skip all of the trig for grayscale colors (z == 0 or z >= 0xd).  Only
the luminance value affects the result, all other terms are zero.
Factor out the conversion to 8-bit integer so that it only has to be
done once for grayscale colors.

For the remainder, calculate the sine and cosine terms only once per
loop iteration.  The hue angle does not need to be clipped to [0, 360),
and the offsets from it can be calculated instead of stored in a table.