Reduce allocation for translation table
Rename TBL_BASE and TBL_BASE2 to tree and data, respectively.

Split ALLOC_SIZE into TREE_SIZE and BLOCK_SIZE.  TREE_SIZE is now
defined in terms of the number of blocks, and BLOCK_SIZE corresponds to
most of the literal '1024' values.

The allocation for data depends on the total number of instruction
bytes, which doesn't necessarily need to match the tree size, so it now
uses DATA_SIZE.

memory_error() is now inlined so that __LINE__ is closer to the actual
error location, and so it can reference the correct symbol to update.
Manually specifying the size is a hack, and in the future I'd like for
it to automatically reallocate.