Pointer/integer conversion improvements
Before relocation, each block is 256 pointers (to other blocks or to the
data).  After relocation, these pointers are converted to integer
offsets, and the encoding and decoding disagree about the signedness.
Use uintptr_t for both, though the offsets don't necessarily need the
same range as a pointer.

Cast to uintptr_t during pointer alignment manipulations.

INT_MAP is an array of pointers to x86 instructions, not an array of
integers.  Correct its type.