Avoid some of the fixed mappings
Choosing the memory location using MAP_FIXED isn't really portable.
We no longer care about the actual location of either mapping since the
parent commit.  Before that, signed integer underflow could occur
during relocation depending on their relative location.  For example if
data is positioned starting at 0x80000000, tree pointers into it are
interpreted as large negative integers, and subtracting the delta may
underflow.

The initial fseek is unnecessary.  The following ftruncate is necessary
to establish the file size so that modifications to that portion of
memory will be written to the file.

Using mmap is overkill and I may remove it later.