Simplify the table.x86 parser
fbf is set and then cleared in the same loop iteration, so no chance of
seeing a third hex digit while it is set.

lnf followed by a non-hex digit is always caught by the hex digit
decoder, so we don't need to complain later if lnf is set when we
receive some other character.

A hex digit is now required following the '/' that sets amf.  But we
want to allow whitespace here (or anywhere, except between a pair of hex
digits when lnf is set), so specifically ignore it at the start.

A source byte pattern can no longer begin with '/' or ','.

'/' is now ignored when dmf is set (inside '[' and ']').

A new pattern may begin on the same line following the ending '/'.

Remaining changes are some reorganization to better use 'else if' and to
avoid initializing static variables to nonzero values.