Deduplicate the search tree nodes
Reduces the number of nodes from 4662 to 102.  Each node is 1 KB, this
saves approximately 4.5 MB in the executable for the cost of a few more
seconds at compile time.

Finding duplicates is fast, but the main source of slowness is patching
up the pointers when removing a duplicate.  There may be faster methods.

The main source of duplicates are patterns such as:
09 00/00 D0,4: <sequence1> /
09 00 D0,4: <sequence2> /
where the first pattern creates 256 identical nodes for when the 2nd
byte is wild, then the second pattern changes only one of those nodes.
The other 255 can be reduced to a single node, but only after we've
processed all patterns (and ruled out the possibility that some but not
all of the 255 could change).