Don't re-render on Expose or MapNotify events
Instead do nothing and let the next frame paint the damaged area.
Except if we're paused, then it might be an unacceptably long time until
the next frame, so set the needsredraw flag so this will be handled
appropriately.  Using this flag avoids the need to synthesize Expose
events from MapNotify events, which itself is unnecessary because
mapping the window will generate both events.