Improve waiting for ShmCompletion
The old method of setting ev.type to -1 was flawed.  If any other type
of event is dequeued first, ev.type gets clobbered for the next
iteration, and we forget that we're waiting.  Use a different flag
variable instead.

The purpose of waiting is so that we don't modify the shared memory,
which we weren't preventing since it may have tried to re-render before
receiving the ShmCompletion.  Now the X input loop should block until
ShmCompletion is received, and there are no attempts to re-render until
after that loop finishes.  The next re-render occurs at the next frame,
or after the next frame time if paused.  Beware, each RenderImage must
be followed by a ShmCompletion; the two proceed in lockstep.