Update event subscriptions
Configure the window's input mask before mapping the window (so that we
do get the initial MapNotify, if the order actually matters), and leave
out the focus change events unless we're going to use them.  Sort the
event handler arms to roughly match the mask order.

The root window is already mapped, so we don't get any event due to the
XMapWindow call.