Split rendering into two steps
The framebuffer only needs to be drawn once per frame, but may need to
be sent to the server multiple times (particularly when paused) in
response to Expose events.  Do a bit of renaming to better distinguish
the two operations (redraw vs. refresh).

Always do the redraw steps once per frame, unless the frame is being
skipped.  This ensures the framebuffer is up-to-date and can be sent to
the server whenever needed, since the drawing code is somewhat stateful.