Combine the frameskip and desync logic
This should make the purpose of the modulo-20 more clear, that we want
to draw every 20th frame even if we're behind, and desync if we are even
further behind.  But if we're continually desync'd, the frame counter
will be advancing by some random increment depending on the current
time, and it should draw more often than every 20th frame.

Any delay accumulated in the input loop will now be attributed to the
next cycle.