Update trace printing
It appears to be trying to access the registers stacked by the `pusha`
instruction just before the call, in particular %ebx at (&s)[4] with the
6502 address to decode.

On my machine it's copying `s` to 16 bytes below the return address
pushed via the call instruction, then all indexing is relative to that
location, so the registers are at (&s)[12] through (&s)[5] inclusive.

The proper way to do this is to either specify each register in the
parameter list, or use varargs.  But varargs would require a dummy named
parameter.