Add remaining branches
Fix the cycle count for backwards branches.  We get a page crossing if
the relative offset is positive with a carry out of bit 7, or negative
and there is no carry (a borrow) out of bit 7.  Only the first case was
handled properly.  We need to XOR the carry out of bit 7 with the sign
of the relative offset to find if a page boundary was crossed.