Switch read/write indexing to post-increment
Removes all of the small modifications to the base address (_RAM) for
any memory operation.  Rewrite all the immediate variants of
instructions to load the address of the immediate into %ebx and
(assuming the immediate is in RAM) run the same code as the zero-page
variants.

Handle 16-bit PC overflow correctly by incrementing %di, not %edi.