Move zero/sign-extension to assembly that requires it
The return value for input() is sign-extended only because the N flag is
stored separately in %dh, so make it the assembly code's responsibility.

The 'val' parameter for output() and all the mapper functions is usually
(but not always) masked with 0xff because the assembly code isn't
consistent in what gets returned in val's high bits.  The parameter type
is now unsigned char, but since the calling convention isn't
particularly clear whether it's the caller's or callee's responsibility
to zero-extend, ensure we do it at least once in the caller.